+++ ui/src/shared/components/Dygraph.js
@@ -367,9 +367,13 @@ export default class Dygraph extends Component {
     // Disallow screen overflow of legend
     const isLegendBottomClipped = graphBottom + legendHeight > screenHeight
 
-    const legendTop = isLegendBottomClipped
-      ? graphHeight + 8 - legendHeight
-      : graphHeight + 8
+    const legendTop = isLegendBottomClipped ? -legendHeight : graphHeight + 8
+
+    if (isLegendBottomClipped) {
+      this.legendRef.classList.add('dygraph-legend--above')
+    } else {
+      this.legendRef.classList.remove('dygraph-legend--above')
+    }
 
     this.legendRef.style.left = `${legendLeft}px`
     this.legendRef.style.top = `${legendTop}px`